+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 FDSTools Changelog
 ==================
+### Version 2.0.4
+- Microhaplotype alleles with SNPs adjacent to any of the microhaplotype
+  positions are now recognised.
+
+
 ### Version 2.0.3
 - Added NimaGen IDseek OmniSTR built-in library file (ID-OmniSTR).
 - Updated dependency requirement (strnaming~=1.1.3).